created comms and videoretrieve files and workd on heading lock button and text box
made heading lock button and text box work
now for heading lock to work we need to implement the head_PID class correctly.
RIght now it's functionally exactly the same as last year.
