+++ GUI/packages/closed_loop/PIDs.py
@@ -8,8 +8,8 @@ import time
 class altitude_PID:
     def __init__(self, desiredAltitude):
         #initialize PID object
-        desiredAltitude = float(desiredAltitude)
-        self.pid_instance = PID(8, 4, 4, setpoint = desiredAltitude)
+        self.desiredAltitude = float(desiredAltitude)
+        self.pid_instance = PID(8, 4, 4, setpoint = self.desiredAltitude)
         self.pid_instance.sample_time = 0.05
         self.pid_instance.output_limits = (-34.32, 44.32)
         self.force_to_joystick_ratio = 1/34.32
@@ -19,13 +19,15 @@ class altitude_PID:
         output = self.pid_instance(currentValue)
         output = output * self.force_to_joystick_ratio
         return output
+    def getDesiredValue(self):
+        return self.desiredAltitude
 
 
 class depth_PID:
     def __init__(self, desiredDepth):
         self.vertical_thruster_amount = 4
-        desiredDepth = float(desiredDepth)
-        self.pid_instance = PID(-7, -1, -3, setpoint=desiredDepth)
+        self.desiredDepth = float(desiredDepth)
+        self.pid_instance = PID(-7, -1, -3, setpoint=self.desiredDepth)
         self.pid_instance.sample_time = 0.05
         self.pid_instance.output_limits = (-34.32*self.vertical_thruster_amount, 44.13*self.vertical_thruster_amount)
         self.force_to_joystick_ratio_neg = 1/(34.32*self.vertical_thruster_amount)
@@ -37,15 +39,17 @@ class depth_PID:
         output = self.pid_instance(depth_input)
         output = output * self.force_to_joystick_ratio_pos
         return output
+    def getDesiredValue(self):
+        return self.desiredDepth
 
 
 class head_PID:
     def __init__(self, desiredHead):
-        desiredHead = float(desiredHead)
+        self.desiredHead = float(desiredHead)
         self.vector_thruster_amount = 4
         self.vector_thrust_contribution = 0.5
         self.max_moment = 156.9*self.vector_thrust_contribution*self.vector_thruster_amount
-        self.pid_instance = PID(4, 1, 1, setpoint=desiredHead)
+        self.pid_instance = PID(4, 1, 1, setpoint=self.desiredHead)
         self.pid_instance.sample_time = 0.05
         self.pid_instance.output_limits = (-self.max_moment, self.max_moment)
         self.force_to_joystick_ratio = 1/self.max_moment
@@ -63,3 +67,5 @@ class head_PID:
                 output = self.pid_instance(higher_input_value)
         output = output * self.force_to_joystick_ratio
         return output
+    def getDesiredValue(self):
+        return self.desiredHead
